vkd3d: Refactor out various app-detection logic to its own file.
It's getting too unwieldy to have it all rammed into device.c.
