+++ libs/vkd3d/meson.build
@@ -74,6 +74,7 @@ vkd3d_src = [
   'command_queue_vkd3d_ext.c',
   'd3dkmt.c',
   'device.c',
+  'device_workarounds.c',
   'device_vkd3d_ext.c',
   'heap.c',
   'memory.c',
